refactor the get CSRF token
