+++ resources/js/stores/espelhoStore.ts
@@ -1,7 +1,7 @@
 import { defineStore } from "pinia";
 import {reactive, ref} from "vue";
 import {Espelho} from "@/Interfaces/Espelho/Espelho";
-import {useUpdateDataStore} from "@/Stores/updateDataStore";
+import {useUpdateDataStore} from "@/stores/updateDataStore";
 
 export const useEspelhoStore = defineStore("espelho", () => {
