+++ resources/js/stores/promotoria.ts
